requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pandas>=1.5.0
numpy>=1.21.0
Pillow>=9.0.0
scikit-learn>=1.2.0
tqdm>=4.64.0
matplotlib>=3.6.0
jupyter>=1.0.0
ipykernel>=6.0.0